replace regular KONECTReader by FastKONECT / exposing new reader to Python
