+++ networkit/cpp/io/KONECTGraphReader.h
@@ -1,57 +1,43 @@
 /*
- * KONECTReader.h
+ * KONECTGraphReader.h
  *
+ *  Created on: 11.05.2018
+ *      Author: Roman Bange
  */
 
 #ifndef KONECTGRAPHREADER_H_
 #define KONECTGRAPHREADER_H_
 
-#include <fstream>
-#include <iostream>
-#include <string>
 #include <unordered_map>
 
-
+#include "../graph/Graph.h"
 #include "GraphReader.h"
 
 namespace NetworKit {
+  class KONECTGraphReader : public NetworKit::GraphReader{
 
-/*
- * KONECTGraphReader.cpp
- * 
- * Reader for the KONECT graph format, 
- * based on the EdgeListReader.
- * 
- * The KONECT format is described in detail in 
- * http://konect.uni-koblenz.de/downloads/konect-handbook.pdf
- */
-class KONECTGraphReader: public NetworKit::GraphReader {
+	public:
+		/*
+		* If the input graph has multiple edges, you can specify on how these edges are handled.
+		* Keep in mind that NetworKit node id's start with 0 while most KONECT graphs start with 1.
+		* See GraphReader.h for a closer description of the paramters.
+		*
+		* @param[in]	remapNodes	specifies whether node ids should be remapped if non consecutive
+		* @param[in]	handlingmethod	specifies how multiple edges should be handled (only relevant if graph with multiple edges is given)
+		*/
+		KONECTGraphReader(bool remapNodes = false, MultipleEdgesHandling handlingmethod = DISCARD_EDGES);
 
-public:
-	/**
-	 * @param[in]	ignoreLoops	ignores loops in the input graph file, if set to true
-	 * @param[in]	separator	character used to separate values of a line
-	 */
-	KONECTGraphReader(char separator=' ', bool ignoreLoops=false);
+		/**
+		 * Given the path of an input file, read the graph contained.
+		 *
+		 * @param[in]	path	input file path
+		 */
+		virtual Graph read(const std::string& path) override;
 
-	/**
-	 * Given the path of an input file, read the graph contained.
-	 *
-	 * @param[in]	path	input file path
-	 */
-	Graph read(const std::string& path);
+	protected:
 
-
-protected:
-	char separator; 	//!< character separating nodes in an edge line
-	std::string commentPrefix;
-	node firstNode;
-	bool ignoreLoops;
-
-private:
-	Graph readContinuous(const std::string& path);
-
-};
-
+		bool remapNodes;
+		MultipleEdgesHandling multipleEdgesHandlingMethod;
+ 	};
 } /* namespace NetworKit */
 #endif /* KONECTGRAPHREADER_H_ */
