+++ networkit/graphio.py
@@ -3,6 +3,7 @@ from _NetworKit import (METISGraphReader, METISGraphWriter, DotGraphWriter, Edge
 						 GMLGraphWriter, LineFileReader, SNAPGraphWriter, DGSWriter, GraphToolBinaryWriter, GraphToolBinaryReader, \
 						  DGSStreamParser, GraphUpdater, SNAPEdgeListPartitionReader, SNAPGraphReader, EdgeListReader, CoverReader, CoverWriter, EdgeListCoverReader, KONECTGraphReader, GMLGraphReader, ThrillGraphBinaryReader, ThrillGraphBinaryWriter)
 from _NetworKit import Graph as __Graph
+from _NetworKit import _MultipleEdgesHandling as MultipleEdgesHandling
 # local imports
 from .GraphMLIO import GraphMLReader, GraphMLWriter
 from .GEXFIO import GEXFReader, GEXFWriter
@@ -96,7 +97,7 @@ def getReader(fileformat, **kwargs):
 			Format.EdgeListTabOne:		EdgeListReader('\t',1),
 			Format.EdgeListTabZero:		EdgeListReader('\t',0),
 			Format.LFR:			EdgeListReader('\t',1),
-			Format.KONECT:			KONECTGraphReader(' '),
+			Format.KONECT:			KONECTGraphReader(),
 			Format.GML:			GMLGraphReader(),
 			Format.GraphToolBinary:		GraphToolBinaryReader(),
 			Format.MAT:			MatReader(),
